src/crawler/extraction/page_extractor.py
"""
HTML parsing and text content extraction from web pages.
"""

from typing import Dict, Optional, List
from bs4 import BeautifulSoup


class PageExtractor:
    """Extracts structured content from HTML pages."""
    
    def __init__(self, config: Dict):
        """Initialize extractor with configuration."""
        pass
    
    def extract_content(self, html: str, url: str) -> Dict:
        """Extract structured content from HTML."""
        pass
    
    def extract_title(self, soup: BeautifulSoup) -> Optional[str]:
        """Extract page title from HTML."""
        pass
    
    def extract_main_content(self, soup: BeautifulSoup) -> Optional[str]:
        """Extract main text content, removing navigation and sidebars."""
        pass
    
    def extract_links(self, soup: BeautifulSoup, base_url: str) -> List[str]:
        """Extract all links from page."""
        pass
    
    def extract_categories(self, soup: BeautifulSoup) -> List[str]:
        """Extract page categories/tags."""
        pass
    
    def extract_infobox_data(self, soup: BeautifulSoup) -> Dict:
        """Extract infobox data if present."""
        pass
    
    def is_disambiguation_page(self, soup: BeautifulSoup) -> bool:
        """Check if page is a disambiguation page."""
        pass
    
    def get_page_type(self, soup: BeautifulSoup, url: str) -> str:
        """Determine page type (character, location, event, etc.)."""
        pass
    
    def _clean_text(self, text: str) -> str:
        """Clean and normalize extracted text."""
        pass
    
    def _remove_wiki_markup(self, text: str) -> str:
        """Remove wiki-specific markup from text."""
        pass